+++ run.py
@@ -5,4 +5,8 @@ app = create_app()
 
 
 if __name__ == "__main__":
-    app.run()
+    app.run(
+        host=app.config["SERVER_HOST"],
+        port=app.config["SERVER_PORT"],
+        debug=app.config["DEBUG"],
+    )